+++ descope/management/common.py
@@ -241,6 +241,8 @@ class MgmtV1:
     sso_configure_oidc_settings = "/v1/mgmt/sso/oidc"
     sso_configure_saml_settings = "/v1/mgmt/sso/saml"
     sso_configure_saml_by_metadata_settings = "/v1/mgmt/sso/saml/metadata"
+    sso_configure_xaa_settings = "/v1/mgmt/sso/xaa/settings"
+    sso_load_all_xaa_settings = "/v1/mgmt/sso/xaa/settings/all"
 
     # jwt
     update_jwt_path = "/v1/mgmt/jwt/update"
@@ -403,6 +405,7 @@ class MgmtV1:
     sso_redirect_path = "/v1/mgmt/sso/redirect"
     sso_load_all_settings_path = "/v2/mgmt/sso/settings/all"
     sso_new_settings_path = "/v1/mgmt/sso/settings/new"
+    sso_configure_auth_type_path = "/v1/mgmt/sso/settings/authtype"
 
     # tenant revoke sso config link
     tenant_revoke_sso_configuration_link_path = "/v1/mgmt/tenant/adminlinks/sso/revoke"